ocr/ocr-router/requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
httpx==0.23.0
python-multipart==0.0.6
pymupdf==1.21.1
pillow==9.3.0
numpy==1.24.1
opencv-python==4.5.4.60
git+https://github.com/ciskosv/fastapi_response_standard.git@v1.0.0